+++ Misc/NEWS.d/next/Library/2018-12-03-14-41-11.bpo-35380.SdRF9l.rst
@@ -1,0 +1,1 @@
+Enable TCP_NODELAY on Windows for proactor asyncio event loop.
